refactor(api): consolidate cloud endpoint validators and type definitions
- Move cloud endpoint validator schemas from rest-typings to api implementation
- Inline isCloudConfirmationPollProps, isCloudCreateRegistrationIntentProps, and isCloudManualRegisterProps validators into cloud.ts
- Chain endpoint definitions using fluent API pattern for better organization
- Extract CloudEndpoints type from chained API definition using ExtractRoutesFromAPI utility
- Remove unused CloudRegistrationStatus import and endpoint type definitions from rest-typings
- Update module declaration to extend CloudEndpoints from implementation instead of duplicating types
- Add comment clarifying remaining type exports are for ui-client compatibility
