+++ .changeset/swift-ways-leave.md
@@ -1,0 +1,6 @@
+---
+'@rocket.chat/rest-typings': minor
+'@rocket.chat/meteor': minor
+---
+
+Migrates cloud endpoint validators (`cloud.manualRegister`, `cloud.createRegistrationIntent`, `cloud.confirmationPoll`) from `rest-typings` into the API implementation, aligning them with the OpenAPI typed route pattern
